refactor the unit test
